+++ ll/linkList_test.go
@@ -15,12 +15,12 @@ type (
 	}
 )
 
-func testFun() bool {
+func testFun[T any](l *LinkList[T]) bool {
 	return true
 }
 
 func TestLinkList(t *testing.T) {
-	cases := [7]TCase{
+	cases := [11]TCase{
 		{
 			name:  "Add:0",
 			input: 1,
@@ -63,43 +63,77 @@ func TestLinkList(t *testing.T) {
 			iter:  2,
 			want:  []int{2, 2, 3, 3, 3},
 		},
+		{
+			name:  "Displace:0,2",
+			input: 0,
+			iter:  1,
+			want:  []int{3, 2, 2, 3, 3},
+		},
+		{
+			name:  "Displace:1,3",
+			input: 0,
+			iter:  3,
+			want:  []int{3, 3, 2, 2, 3},
+		},
+		{
+			name:  "AddLast",
+			input: 1,
+			iter:  1,
+			want:  []int{3, 3, 2, 2, 3, 1},
+		},
+		{
+			name:  "Displace:1,5",
+			input: 0,
+			iter:  3,
+			want:  []int{3, 1, 2, 2, 3, 3},
+		},
 	}
 
-	ll := New[int](map[string]interface{}{"test": testFun})
+	l, err := New[int](map[string]any{"test": testFun[int]})
+	if err != nil {
+		t.Errorf("on calling `New` function error happened:\nerr => %s", err.Error())
+	}
 
 	for k, c := range cases {
 		if c.name == "AddFirst" {
 			for i := 0; i < c.iter; i++ {
 				temp := c.input
-				ll.AddFirst(&temp)
+				l.AddFirst(&temp)
 			}
 		} else if c.name == "RemoveFirst" {
 			for i := 0; i < c.iter; i++ {
-				ll.RemoveFirst()
+				l.RemoveFirst()
 			}
 		} else if c.name == "AddLast" {
 			for i := 0; i < c.iter; i++ {
 				temp := c.input
-				ll.AddLast(&temp)
+				l.AddLast(&temp)
 			}
 		} else if c.name == "RemoveLast" {
 			for i := 0; i < c.iter; i++ {
-				ll.RemoveLast()
+				l.RemoveLast()
 			}
 		} else if strings.Contains(c.name, "Remove:") {
 			num, _ := strconv.Atoi(c.name[7:])
 			for i := 0; i < c.iter; i++ {
-				ll.Remove(num)
+				l.Remove(num)
 			}
 		} else if strings.Contains(c.name, "Add:") {
 			num, _ := strconv.Atoi(c.name[4:])
 			for i := 0; i < c.iter; i++ {
 				temp := c.input
-				ll.Add(num, &temp)
+				l.Add(num, &temp)
+			}
+		} else if strings.Contains(c.name, "Displace:") {
+			numsSplit := strings.Split(c.name[9:], ",")
+			num1, _ := strconv.Atoi(numsSplit[0])
+			num2, _ := strconv.Atoi(numsSplit[1])
+			for i := 0; i < c.iter; i++ {
+				l.DisplaceIndex(num1, num2)
 			}
 		}
 
-		slice := ll.ReturnAsSlice()
+		slice := l.ReturnAsSlice()
 
 		if len(c.want) != len(slice) {
 			t.Errorf("(%s, %d) => want = %v, got = %v", c.name, k, c.want, slice)
@@ -114,7 +148,7 @@ func TestLinkList(t *testing.T) {
 		}
 	}
 
-	if res, err := ll.Run("test"); res == nil || err != nil {
-		t.Errorf("on calling test something happened:\nerr => %s", err)
+	if res, err := l.Run("test"); res == nil || err != nil {
+		t.Errorf("on calling `Run` for `test` function error happened:\nerr => %s", err.Error())
 	}
 }