+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: first input check in `New` function for functions
+utilities is now in ll package
+test cases updated
 * ✅ test(ll): linklist structure test cases added
 * 🎉 feat(ll): linklist structure added